Add CPI method for native disk update
Certain cloud providers offer the capability to modify disk attributes like
IOPS and throughput directly, eliminating the need to create a new disk. To
accommodate this, this change introduces a new CPI method called update_disk.
Previously, adjusting disk properties required creating a new disk and
transferring the data, a process that could be time-consuming for larger disks.
Should the CPI not support this new method, BOSH will revert to the traditional
approach of disk creation.

Co-authored-by: MSSedusch <sedusch@microsoft.com>